feat: Complete litebike proxy server with universal protocol detection
- Enable full feature set (auto-discovery, DoH, UPnP) for maximum edge case coverage
- Fix all compilation errors and build successfully
- Add comprehensive protocol handlers for HTTP, SOCKS5, TLS, DoH, PAC/WPAD, Bonjour, UPnP
- Implement universal port 8888 with intelligent protocol detection using Patricia Trie
- Update proxy-bridge script for litebike integration with safe SSH commands
- Add extensive test suite and benchmarks for protocol detection
- Support ARM64 SIMD optimizations for mobile devices

+++ src/simple_torture_test.rs
@@ -263,7 +263,7 @@ impl SimpleTortureTest {
                 protocol,
                 bytes_consumed: bytes,
                 processing_time: duration,
-                memory_efficient: duration < Duration::from_millis(size / 1000 + 10), // Rough heuristic
+                memory_efficient: duration < Duration::from_millis((size / 1000 + 10) as u64), // Rough heuristic
             });
 
             if duration > Duration::from_millis(100) {
